Email: Create a new store for the new add account
If the user add the same account which is deleted just a moment ago, the
new hostAuth object will be equal to the old hostAuth which belongs to
the deleted account. Then we will got the old store for the new account,
but actually it is wrong.

So we need judge if the got store's id is same as the given account's id.
If they are different, it means it does not belong to the new account.
Then create a new one for the new account.

CRs-Fixed: 354221

Change-Id: Ic1a8d6d76114e0a0d975afa1bc95c7a4dd982729