Email: Add the attachment's query columns to matched the OpenableColumns.
The attachment's query columns didn't contain the OpenableColumns. If the
user try to view or open this attachment, the choosed application couldn't
get the display name and size when they use the default OpenableColumns api
to read the data saved from cursor.

So add these two columns to query projection as default which matched the
OpenableColumns defined.

CRs-Fixed: 385446

Change-Id: I935eb778e0ff4f14934274dbd61a219afb71cc58