Email: Update the folders' name when system language change
The folder names of "inbox" "draft" and "sent" was initialized when the
FolderProperties class was created. After device language changes, the
instance will not be recreated. As a result, the folders display name will
not change.

To solve this problem, add logic in FolderProperties to reload the special
folder names when device locale has been changed.

CRs-Fixed: 371600

Change-Id: I625244cc928758e51367f4f305dffbba852dd28f