Email: Add play option for audio and video files in attachment
If the email with ogg file attatched was sent by Google mail server, the
mime type of the ogg file will be application/octet-stream, email app will
infer the mime type to application/ogg, which is pretty smart and good.
However, when the attachment tab displayed, only mime type started with
"video" will be treated as can be played. That caused the ogg file could
not be played in attachment.

Check the file type, if the file is audio or video, treat it as can be
played.

CRs-Fixed: 405080

Change-Id: I3da117a3f609ee18229ea3c3d988507b014fb22d