Email: Resolve the draft maybe restored with the source message.
If the draft has the source message such as the forward mail, when the user
try to view or edit this draft, the content will be restored with the source
message. This is caused by it will be restored when it got the action (Such
as forward, reply, and so on.) changed event.

So need to get if it need restore the content on action changed.

CRs-Fixed: 415044

Change-Id: Iba5d1134895e29cda1c56cf1c3ca76a8cf0f57cc