Email: Resolve the selected number may be incorrect problem.
Once the message list goes into the selected mode by selected some mails in,
the list, and then modify one of the selected item, for example, delete one
selected mail. Then back to the message list. The selected number will be not
updated, and it will be displayed as the last selected number. This is caused
by the selected set will be not updated with the new list.

So need to update the selected set according to the new list items.

CRs-Fixed: 424529

Change-Id: I94d8e0dbc28d5fceb4f9d5fed7556bc625fd464d