Email: Change intent name to recive intents from bluetooth MAP
Change intent name in email app, to name of the intent which are
being broadcasted from bluetooth map app, to support bluetooth MAP
Profile for email instance.

Add change to sync the folder at which delete operation is
performed on the message, so that the UI shows updated list
after deleting the message.

CRs-fixed: 594604,594565

Change-Id: Ib27951ca61ab367d7b07a86800339e2c2ed6a688