Email: Update the last success sync time for mailbox.
As the mailbox maybe didn't contain any mail, and it will never update the
database. And the mail list couldn't get the load finished callback. So the
load status couldn't be update even the sync process is finished.

Now we will update the last success sync time for this mailbox and it will
notify the mail list data changed.

CRs-fixed: 629445

Change-Id: I607fc7bcb8ccbae0119b43cb25e635db2f6ee3d6