Email: Hide the feedback button if nowhere to feedback
The string value of email_feedback_uri in strings.xml presents where
to submit the feedback. If it is empty or invalid, shall hide the
feed back button.

Add the switch to check email_feedback_uri to show feedback button for
the Email module. Add the switch to set the capability for UnifiedEmail
to use.

Change-Id: I7e790c67bfdf2b3096ccb9ce14e120b54e8e13d9
CRs-fixed: 737872