Email: Double check the queue if empty before the service stop or wait.
If the attachment service already finished the process queue action, even the
queue add some request again, the request can not handle. It is caused by the
service will stopped or wait 30 minutes after process queue.

So before the service stopped or wait, we'd like to double check the queue if
empty, and after wait 5 seconds, check again.

CRs-fixed: 749621

Change-Id: I336c83211ed05652fea74ba2be9e10c358b54076