Update the mail's attachment flag as false if it do not contains any atts.
When we send the message and mount the SD card, we can't find the
attachment in the device, so the device can't be added to the provider,
attachment in the device, so the device can't be added to the provider,
but the flag of the attachment is not delete from the provider, so it but
the flag of the attachment is not delete from the provider, so it will
show the attachment icon on the message. will show the attachment icon on
the message.

Change the attachment flag in the provider when the attachment can't Change
the attachment flag in the provider when the attachment can't 14 be found.

CRs-Fixed: 768290

Change-Id: I4fe9f76a10ea5c11e217c2c81aa9112b3ce9f37d