Email: Fix exit after rotate in account setup UI
After rotate the screen, the mResultCode of Activity is reset to
default value 'RESULT_CANCELED'. if tap next button to start the
MailActivity, the result_code will be checked, MailActivity finish
if the result_code is RESULT_CANCELED.

So we save the state of account loading task and then save in the
savedInstance to resotre it when rotate the screen. In the MailActivity
recreate process the result_code will set to the right value
according to the state.

Change-Id: I3f153892ee45c31d660bfbf36885151c1b0fa7de
CRs-Fixed: 2203811